+++ replay-control-app/tests/common/mod.rs
@@ -160,6 +160,7 @@ pub fn register_server_fns() {
     server_fn::axum::register_explicit::<server_fns::RescanGameLibrary>();
     server_fn::axum::register_explicit::<server_fns::GetSetupStatus>();
     server_fn::axum::register_explicit::<server_fns::DismissSetup>();
+    server_fn::axum::register_explicit::<server_fns::GetMetadataLibraryOverview>();
     server_fn::axum::register_explicit::<server_fns::GetLocalManuals>();
     server_fn::axum::register_explicit::<server_fns::DownloadManual>();
     server_fn::axum::register_explicit::<server_fns::DeleteManual>();